Move actions and reducers on specific folder and change `app` reducer to `auth`
