+++ src/actions/auth.js
@@ -1,11 +1,11 @@
 import fetch from 'isomorphic-fetch';
 import { createAction } from 'redux-actions';
 
-export const AUTHENTICATE_REQUEST = '@app/AUTHENTICATE_REQUEST';
+export const AUTHENTICATE_REQUEST = '@auth/AUTHENTICATE_REQUEST';
 export const authenticateRequest = createAction(AUTHENTICATE_REQUEST);
-export const AUTHENTICATE_SUCCESS = '@app/AUTHENTICATE_SUCCESS';
+export const AUTHENTICATE_SUCCESS = '@auth/AUTHENTICATE_SUCCESS';
 export const authenticateSuccess = createAction(AUTHENTICATE_SUCCESS);
-export const AUTHENTICATE_FAILURE = '@app/AUTHENTICATE_FAILURE';
+export const AUTHENTICATE_FAILURE = '@auth/AUTHENTICATE_FAILURE';
 export const authenticateFailure = createAction(AUTHENTICATE_FAILURE);
 
 export const authenticate = () =>
