+++ src/app.js
@@ -5,11 +5,11 @@ export default class App extends Component {
   render() {
     return (
       <div>
-        <Header username={this.props.app.user.name} />
+        <Header username={this.props.auth.user.name} />
         <div>
           {React.cloneElement(
             this.props.children,
-            { app: this.props.app }
+            { auth: this.props.auth }
           )}
         </div>
       </div>
